completed basic SGD implementation
This implementation is a lot more flexible, thanks to the use of interfaces. There are a lot of parameters for SGD, so I decided to use a struct to store them rather than using function arguments.

The biggest upgrade from my previous SGD package is that the learning rate can automatically adjust based on the epoch.