drafted anyrnn types and mapping
You'll notice that map.go is 85 lines, whereas the old weakai/rnn/block_seq_func.go is 303 lines, yet they do the same thing. That is for two reasons:

 1. No explicit R-operator support
 2. More elegant solution to differing sequence len

Using PresentMap in combination with Expand() and Reduce(), we provide a natural, built in way to deal with variable-length sequences. The result is that the code is MUCH simpler.