better error system for sample lists
It was kind of silly that you'd have to implement your own Fetcher if you wanted to be able to return sample-fetch errors. Take imagenet for example: it is perfectly reasonable to use anyff, but you want the GetSample() method on your SampleList to do an operation which might fail (read a file).