pool to anyvec64 in anyctc
When using the CUDA anyvec32 implementation, ctc is *extremely* slow. Basically, since we do a ton of component-wise operations, it makes the most sense to use a Creator that we actually control (and which lives on the CPU). I expect this to boots performance of CTC when using cuda by at least a factor of 20, just by not using cuda anymore.