added Markov RNN block
This has a number of different use cases. For example, you could use it in an RL setting to pass the last few frames to a convolutional neural network. You could also use it as the first layer in an NLP network to give each timestep a little more structured context.